Moved the DetermineLinkEndPoints function from linkRelatedSwitchingOperationUtility to PrepareForwardingAutomation
